% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourPL_functions.R
\name{FourPL1}
\alias{FourPL1}
\alias{FourPL1_fixa}
\alias{FourPL1_fixy0}
\alias{FourPL1_fixa_fixy0}
\alias{FourPL1_deriv}
\title{Four parameter logistic (with explicit inflection point)}
\usage{
FourPL1(x, a, bta, c, y0)

FourPL1_fixa(x, a = 100, bta, c, y0)

FourPL1_fixy0(x, a, bta, c)

FourPL1_fixa_fixy0(x, a = 100, bta, c)

FourPL1_deriv(x, y0, a, b, c)
}
\arguments{
\item{x}{The explanatory/independent variable value.}

\item{a}{Parameter \mjseqn{a}.}

\item{bta}{Parameter \mjseqn{\beta} where \mjseqn{b = e^{\beta}}.}

\item{c}{Parameter \mjseqn{c}.}

\item{y0}{Parameter \mjseqn{y_{0}}.}

\item{b}{Parameter \mjseqn{b}.}
}
\value{
The calculated response/dependent value value.
}
\description{
To be used by
\code{\link[dormancymetrics]{FourPLfit}}. \loadmathjax
}
