% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourPL_functions.R
\name{FourPL2}
\alias{FourPL2}
\alias{FourPL2_fixa}
\alias{FourPL2_fixy0}
\alias{FourPL2_fixa_fixy0}
\alias{FourPL2_deriv}
\title{Four paramter logistic (with implicit inflection point)}
\usage{
FourPL2(x, a, bta, k, y0)

FourPL2_fixa(x, a = 100, bta, k, y0)

FourPL2_fixy0(x, a, bta, k)

FourPL2_fixa_fixy0(x, a = 100, bta, k)

FourPL2_deriv(x, y0, a, b, k)
}
\arguments{
\item{x}{The explanatory/independent variable value.}

\item{a}{Parameter \mjseqn{a}.}

\item{bta}{Parameter \mjseqn{\beta} where \mjseqn{b = e^{\beta}}.}

\item{k}{Parameter \mjseqn{k}.}

\item{y0}{Parameter \mjseqn{y_{0}}.}

\item{b}{Parameter \mjseqn{b}.}
}
\value{
The calculated response/dependent value value.
}
\description{
To be used by
\code{\link[dormancymetrics]{FourPLfit}}. \loadmathjax
}
