% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourPL_withWE.R
\name{FourPL_withWE}
\alias{FourPL_withWE}
\title{Function/model selection helper with \code{trycatch}}
\usage{
FourPL_withWE(
  data,
  fix.a,
  fix.y0,
  starta,
  startbta,
  startc,
  startk,
  starty0,
  maxiter,
  algorithm = "lm",
  warnOnly = FALSE,
  inflection.point = c("explicit", "implicit"),
  time.scale = c("linear", "log")
)
}
\arguments{
\item{fix.a}{Fix a as the actual maximum germination percentage at the end of
the experiment.}

\item{fix.y0}{Force the intercept of the y axis through 0.}

\item{inflection.point}{Either \code{"explicit"} or \code{"implicit"}. See
\strong{Details}.}

\item{time.scale}{Either \code{"linear"} for fitting a four-parameter
logistic function or \code{"logarithmic"} for fitting a four-parameter
log-logistic function.}
}
\description{
Not exported. Strictly internal
}
\keyword{internal}
