% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourPLfit.R
\name{FourPLfit}
\alias{FourPLfit}
\title{Fit four-parameter logistic or log-logistic function}
\usage{
FourPLfit(
  germ.counts,
  intervals,
  rep,
  total.seeds,
  fix.y0 = TRUE,
  fix.a = TRUE,
  inflection.point = c("explicit", "implicit"),
  time.scale = c("linear", "log"),
  tmax = max(int),
  tries = 3
)
}
\arguments{
\item{germ.counts}{Germination counts at each time interval (days of seed
storage) as a numeric vector.}

\item{intervals}{Time intervals of seed storage as a numeric vector.}

\item{rep}{Optional. An integer vector indicating the replication
corresponding to each germination count.}

\item{total.seeds}{Total number of seeds as an integer vector of unit length.}

\item{fix.y0}{Force the intercept of the y axis through 0.}

\item{fix.a}{Fix a as the actual maximum germination percentage at the end of
the experiment.}

\item{inflection.point}{Either \code{"explicit"} or \code{"implicit"}. See
\strong{Details}.}

\item{time.scale}{Either \code{"linear"} for fitting a four-parameter
logistic function or \code{"logarithmic"} for fitting a four-parameter
log-logistic function.}

\item{tmax}{The time interval up to which AUC is to be computed.}

\item{tries}{The number of tries to be attempted to fit the curve. Default is
3.}
}
\value{
A list with the following components: \item{data}{A data frame with
  the data used for computing the model.} \item{Parameters}{A data frame of
  parameter estimates, standard errors and p value.}  \item{Fit}{A one-row
  data frame with estimates of model fitness such as log likelyhoods, Akaike
  Information Criterion, Bayesian Information Criterion, deviance and
  residual degrees of freedom.}  \item{a}{The upper asymptote, or maximum
  germination percentage.}  \item{b}{The mathematical parameter controlling
  the shape and steepness of the dormancy release curve.} \item{c}{The
  inflection point which represents the time point of storage with maximum
  rate of dormancy release.}  \item{y0}{The lower asymptote, or intercept on
  the y axis.} \item{TMRDR}{Time at maximum rate of dormancy release.}
  \item{AUC}{The estimate of area under the curve.} \item{msg}{The message
  from \code{\link[gslnls]{gsl_nls}}.} \item{isConv}{Logical value
  indicating whether convergence was achieved.} \item{model}{The raw fitted
  model output as a list of class \code{gsl_nls}.}
}
\description{
Fit a four-parameter logistic or log-logistic function
\insertCite{bentsink_natural_2010,bentsink_identification_2011,postma_maternal_2015,fogliatto_effect_2020}{dormancymetrics}
to germination count data recorded at sequential time intervals of seed
storage to model a dormancy release curve and compute the associated
parameters. \loadmathjax
}
\details{
The germination count data of a seed lot recorded at sequential time
intervals of storage can be modelled to fit a \strong{four-parameter
logistic} function defined as follows
\insertCite{bentsink_natural_2010,bentsink_identification_2011}{dormancymetrics}.

\mjsdeqn{y(x) =  y_{0} + \frac{a - y_{0}}{1 + e^{-b(x - c)}}}

Where, \mjseqn{y} is the germination percentage at the \mjseqn{x} time
interval after storage, \mjseqn{y_{0}} is the lower asymptote, or intercept
on the y axis, \mjseqn{a} is the upper asymptote, or maximum germination
percentage, \mjseqn{b} is a mathematical parameter controlling the shape and
steepness of the dormancy release curve (the larger the \mjseqn{b} parameter,
the steeper the rise toward the asymptote \mjseqn{a}, and the shorter the
time between initial germination value and maximum germination), and
\mjseqn{c} is the inflection point which represents the time point of storage
with maximum rate of dormancy release. \mjseqn{c} is the \mjseqn{DSDS_{50}}
(days of seed dry storage required to reach 50\% germination).

This model has the inflection point \mjseqn{c} explicitly included.
Alternatively the data can also be modelled to fit a four-parameter logistic
function with implicit inflection point
\insertCite{postma_maternal_2015}{dormancymetrics} defined as follows.

\mjsdeqn{y(x) =  y_{0} + \frac{a - y_{0}}{1 + k \cdot e^{-b \cdot x}}}

Where, \mjseqn{k} controls the initial value of the function at \mjseqn{x =
0}. It determines the initial steepness and starting value of the curve. It
is the asymmetry or horizontal shift factor which affects where the curve
reaches its inflection point and introduces asymmetry.

\mjsdeqn{k = e^{bc} \; \textrm{and} \; c = \frac{\ln{k}}{b}}

This is useful in cases where more than 50\% germination is achieved in the
first time interval itself.

This model is more suitable where the dormancy release over a linear time
scale is smooth and symmetrical. In cases of rapid early dormancy release
with a long tail of the curve, a \strong{log-logistic model} is more
appropriate \insertCite{fogliatto_effect_2020}{dormancymetrics}. Here the
time \mjseqn{x} is scaled logarithmically.

\mjsdeqn{y(x) =  y_{0} + \frac{a - y_{0}}{1 + e^{-b(\log(x) - \log(c))}}}

\mjsdeqn{y(x) =  y_{0} + \frac{a - y_{0}}{1 + k \cdot x^{b}} \; \left (\textrm{As} \;
e^{b \cdot \log(x)} = x^{b}  \right )}

In \code{FourPLfit}, these models have been reparameterized by substituting
\mjseqn{b} with \mjseqn{e^{\beta}} to constraint \mjseqn{b} to positive
values only.

\mjsdeqn{y(x) =  y_{0} + \frac{a - y_{0}}{1 + e^{-e^{\beta}(x - c)}}}

\mjsdeqn{y(x) = y_0 + \frac{a - y_{0}}{1 + k \cdot e^{-e^{\beta} \cdot x}}}

\mjsdeqn{y(x) =  y_{0} + \frac{a - y_{0}}{1 + e^{-e^{\beta}(\log(x) -
\log(c))}}}

\mjsdeqn{y(x) =  y_{0} + \frac{a - y_{0}}{1 + k \cdot x^{e^{\beta}}}}
}
\examples{
x <- c(2, 1, 2, 2, 0, 0, 2, 2, 0, 2, 2, 0, 2, 2, 2, 6, 8, 10, 8, 19,
       8, 4, 11, 4, 22, 19, 25, 16, 21, 30, 40, 33, 34, 36, 44, 42,
       42, 39, 42, 38, 47, 42, 50, 44, 48, 50)
y <- c(0, 0, 14, 14, 18, 18, 20, 20, 21, 21, 22, 22, 23, 23, 23, 23,
       24, 24, 24, 24, 25, 25, 25, 25, 25, 25, 25, 25, 26, 26, 26, 26,
       26, 26, 26, 26, 27, 27, 27, 27, 27, 27, 27, 27, 27, 27)
rep <- rep(1:2, 23)
int <- rep(c(0, 3, 6, 9, 12, 15, 18, 21, 24, 27, 31, 35, 39, 43, 47, 52,
             57, 62, 67, 72, 82, 92, 102), each = 2)
total.seeds = 50

# Logistic fit
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
FourPLfit(germ.counts = x, intervals = int, rep = rep, total.seeds = 50,
          fix.y0 = TRUE, fix.a = TRUE,
          inflection.point = "explicit", time.scale = "linear")

FourPLfit(germ.counts = x, intervals = int, rep = rep, total.seeds = 50,
          fix.y0 = TRUE, fix.a = TRUE,
          inflection.point = "implicit", time.scale = "linear")

# Log-logistic fit
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
FourPLfit(germ.counts = y, intervals = int, rep = rep, total.seeds = 50,
          fix.y0 = TRUE, fix.a = TRUE,
          inflection.point = "explicit", time.scale = "log")

FourPLfit(germ.counts = y, intervals = int, rep = rep, total.seeds = 50,
          fix.y0 = TRUE, fix.a = TRUE,
          inflection.point = "implicit", time.scale = "log")

}
\references{
\insertAllCited{}
}
