% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourPLL_functions.R
\name{FourPLL2}
\alias{FourPLL2}
\alias{FourPLL2_fixa}
\alias{FourPLL2_fixy0}
\alias{FourPLL2_fixa_fixy0}
\alias{FourPLL2_deriv}
\title{Four paramter log-logistic (with implicit inflection point)}
\usage{
FourPLL2(x, a, bta, k, y0)

FourPLL2_fixa(x, a = 100, bta, k, y0)

FourPLL2_fixy0(x, a, bta, k)

FourPLL2_fixa_fixy0(x, a = 100, bta, k)

FourPLL2_deriv(x, y0, a, b, k)
}
\arguments{
\item{x}{The explanatory/independent variable value.}

\item{a}{Parameter \mjseqn{a}.}

\item{bta}{Parameter \mjseqn{\beta} where \mjseqn{b = e^{\beta}}.}

\item{k}{Parameter \mjseqn{k}.}

\item{y0}{Parameter \mjseqn{y_{0}}.}

\item{b}{Parameter \mjseqn{b}.}
}
\value{
The calculated response/dependent value value.
}
\description{
To be used by
\code{\link[dormancymetrics]{FourPLfit}}. \loadmathjax
}
